requirements.txt
streamlit==1.41.1
pandas==2.2.3
openai==1.61.0
langchain-core==0.3.33
langchain==0.3.17
langchain-openai==0.3.3
langchain-community==0.3.16
langchain-experimental==0.3.4
langgraph==0.2.69
langgraph-checkpoint==2.0.10
matplotlib==3.10.0
seaborn==0.13.2
tabulate
cartopy==0.24.1
requests>=2.32.4
beautifulsoup4==4.12.3
pangaeapy==1.0.20
tqdm==4.67.1
xarray==2025.1.2
gsw==3.6.19
adjustText==1.3.0
pypdf==5.2.0
langchain-text-splitters==0.3.5
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
netCDF4==1.6.5
numpy==1.26.4
langchain_anthropic
pytest==8.3.4
pytest-mock==3.14.0
pytest-asyncio==0.25.3
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
python-dotenv
